pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillcheck"
version = "0.1.0"
description = "Audit, lint, probe, and attest Claude Skills with policy-driven checks."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
  {name = "SKILLCHECK contributors"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Build Tools"
]
dependencies = [
  "typer[all]>=0.9.0",
  "pydantic>=1.10,<3.0",
  "pyyaml>=6.0",
  "rich>=13.0",
  "tabulate>=0.9.0"
]

[project.urls]
Homepage = "https://github.com/example/skillcheck"
Documentation = "https://github.com/example/skillcheck#readme"

[project.scripts]
skillcheck = "skillcheck.cli:main"

[project.optional-dependencies]
otel = [
  "opentelemetry-sdk>=1.22.0",
  "opentelemetry-exporter-otlp>=1.22.0"
]
viz = [
  "matplotlib>=3.7"
]
attest = [
  "sigstore>=2.0.0"
]
dev = [
  "pytest>=7.4",
  "ruff>=0.4.0",
  "mypy>=1.8.0"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
